requirements.txt
click==8.0.3
colorama==0.4.4
decouple==0.0.7
dnspython==2.1.0
Flask==2.0.2
Flask-PyMongo==2.3.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pymongo==3.12.1
python-dotenv==0.19.1
Werkzeug==2.0.2
